backend/requirements.txt
# Core Python libraries (usually included with Python)
# os - built-in
# json - built-in
# glob - built-in
# sys - built-in
# re - built-in
# math - built-in
# io - built-in
# hashlib - built-in
# random - built-in
# typing - built-in

# Document Processing Libraries
PyMuPDF==1.23.26          # fitz - PDF processing
python-docx==1.1.0        # docx - DOCX file processing
python-pptx==0.6.23       # pptx - PowerPoint file processing

# Image Processing and OCR
Pillow==10.1.0             # PIL - Image processing
pytesseract==0.3.10       # OCR for text extraction from images

# HTTP Requests (for OpenAI API)
requests==2.31.0          # HTTP requests for API calls

# Machine Learning and Embeddings
numpy==1.26.2             # Numerical computing (used in test_embeddings.py)
sentence-transformers==5.1.1  # HuggingFace embeddings (optional)

# Optional: Advanced ML libraries (if you want to extend functionality)
# torch==2.1.1             # PyTorch (dependency of sentence-transformers)
# transformers==4.36.0     # HuggingFace transformers
# scikit-learn==1.3.2      # Machine learning utilities

# Development and Testing
# pytest==7.4.3           # Testing framework (optional)
# black==23.11.0           # Code formatting (optional)
# flake8==6.1.0            # Code linting (optional)

# Web Framework (for API endpoints)
fastapi==0.104.1           # Fast web framework
uvicorn==0.24.0            # ASGI server
python-multipart==0.0.6    # For file upload support